infrastructure/testdriver/set_timezone.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver set_timezone method</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
function defaultTimezone() {
  return (new Intl.DateTimeFormat()).resolvedOptions().timeZone;
}
async_test(t => {
  let timezone = "Asia/Taipei";
  test_driver
    .set_timezone(timezone)
    .then(() => {
      assert_equals(defaultTimezone(), timezone);
      t.done()
    })
    .catch(() => assert_unreached("set_timezone failed"));
});
</script>
